Move authentification logic to auth.js module
It's not just for refactoring it allow me to mock the module for testing
